+++ src/forze_fastapi/handlers/__init__.py
@@ -1,5 +1,6 @@
 from .exceptions import register_exception_handlers
+from .logging import register_uvicorn_logging_interceptor
 
 # ----------------------- #
 
-__all__ = ["register_exception_handlers"]
+__all__ = ["register_exception_handlers", "register_uvicorn_logging_interceptor"]